[15]: Expense page (#42)
* expense page reconstruction. part 1

* added identity + new migrations

* added cancellation token & improve logger

* adjusted specification pattern

* improved tests and removed identity

* adjusted exchange dialog

* added filtering + queryMapping

* category & dialog multiples

* add updateRange categories

* add category dialog

* expense refactor

* angular update 19v

* update angular & typescript & component 19.4

* update cdk 19

* update to 19

* update cdk to 20v

* fixed hooks & modified for v20 angular

* fix values

* modified capital

* adjust capital requests & ui

* created & background job execution optimization

* API code fixes & UI refactor

* modified caching & error handling

* fixed category & refactor the code structure and build

* improved UI & removed unnecessary code

* draft changes of update expense approach

* improve update expense logic

* fix validators

* fix tests

* added restore mechanism

+++ src/Deed/Deed.Application/Capitals/Commands/Delete/DeleteCapitalCommandHandler.cs
@@ -13,16 +13,21 @@ internal sealed class DeleteCapitalCommandHandler(
 {
     public async Task<Result> Handle(DeleteCapitalCommand command, CancellationToken cancellationToken)
     {
-        var capital = await repository.GetAsync(new CapitalByIdSpecification(command.Id));
+        var capital = await repository.GetAsync(new CapitalByIdSpecification(command.Id, true, true, true, true)).ConfigureAwait(false);
 
         if (capital is null)
         {
             return Result.Failure(DomainErrors.General.NotFound(nameof(capital)));
         }
 
+        if (capital.HasReferences())
+        {
+            return Result.Failure(DomainErrors.Capital.ReferenceExists);
+        }
+
         repository.Delete(capital);
 
-        await unitOfWork.SaveChangesAsync(cancellationToken);
+        await unitOfWork.SaveChangesAsync(cancellationToken).ConfigureAwait(false);
 
         return Result.Success();
     }
